add staticStyle to useFela for theme-aware styles
