+++ packages/fela/src/cssifyStaticStyle.js
@@ -3,7 +3,7 @@ import { processStyleWithPlugins, STATIC_TYPE } from 'fela-utils'
 
 import minifyCSSString from './minifyCSSString'
 
-export default function cssifyStaticStyle(staticStyle, renderer) {
+export default function cssifyStaticStyle(staticStyle, renderer, props) {
   if (typeof staticStyle === 'string') {
     return minifyCSSString(staticStyle)
   }
@@ -11,7 +11,8 @@ export default function cssifyStaticStyle(staticStyle, renderer) {
   const processedStaticStyle = processStyleWithPlugins(
     renderer,
     staticStyle,
-    STATIC_TYPE
+    STATIC_TYPE,
+    props
   )
 
   return cssifyObject(processedStaticStyle)
