+++ packages/preact-fela/src/useFela.js
@@ -1,31 +1,6 @@
-// $FlowFixMe
 import { useContext } from 'preact/hooks'
-import { combineRules } from 'fela'
+import { useFelaFactory } from 'fela-bindings'
 
-import { RendererContext, ThemeContext } from './context'
+import { ThemeContext, RendererContext } from './context'
 
-export default function useFela(props = {}) {
-  const renderer = useContext(RendererContext)
-  const theme = useContext(ThemeContext) || {}
-
-  if (!renderer) {
-    throw new Error(
-      'The "useFela" hook can only be used  inside a "RendererProvider"'
-    )
-  }
-
-  const propsWithTheme = {
-    ...props,
-    theme,
-  }
-
-  function css(...rules) {
-    return renderer.renderRule(combineRules(...rules), propsWithTheme)
-  }
-
-  return {
-    renderer,
-    theme,
-    css,
-  }
-}
+export default useFelaFactory(RendererContext, ThemeContext, useContext)
