+++ packages/react-fela/src/useFela.js
@@ -1,41 +1,6 @@
 import { useContext } from 'react'
-import { combineRules } from 'fela'
+import { useFelaFactory } from 'fela-bindings'
 
-import { RendererContext, ThemeContext } from './context'
+import { ThemeContext, RendererContext } from './context'
 
-function getPropsWithTheme(props, theme) {
-  if (props) {
-    return {
-      ...props,
-      theme,
-    }
-  }
-
-  return {
-    theme,
-  }
-}
-
-export default function useFela(props) {
-  const renderer = useContext(RendererContext)
-  const theme = useContext(ThemeContext) || {}
-
-  if (!renderer) {
-    throw new Error(
-      'The "useFela" hook can only be used  inside a "RendererProvider"'
-    )
-  }
-
-  // we add the theme to props so that it can be used within styles
-  const propsWithTheme = getPropsWithTheme(props, theme)
-
-  function css(...rules) {
-    return renderer.renderRule(combineRules(...rules), propsWithTheme)
-  }
-
-  return {
-    renderer,
-    theme,
-    css,
-  }
-}
+export default useFelaFactory(RendererContext, ThemeContext, useContext)
