+++ website/docs/latest/api/fela/renderer.mdx
@@ -178,11 +178,11 @@ Renders static styles.
 
 ### Arguments;;renderStatic-arguments
 
-| Argument | Type                   | Description                                                                   |
-| -------- | ---------------------- | ----------------------------------------------------------------------------- |
-| style    | _string_<br />_Object_ | Either a pure CSS string or an object of style declarations.                  |
-| selector | _string_               | If `style` is passed as an object you **must** specify a `selector` selector. |
-| props    | _Object?_              |                                                                               |
+| Argument | Type                   | Description                                                                    |
+| -------- | ---------------------- | ------------------------------------------------------------------------------ |
+| style    | _string_<br />_Object_ | Either a pure CSS string or an object of style declarations.                   |
+| selector | _string_               | If `style` is passed as an object you **must** specify a `selector` selector.  |
+| props    | _Object?_              | An object of props that is passed to the plugins processing the style objects. |
 
 ### Example;;renderStatic-example
